Break constructor apart
Into classes common to all pyiron ontologies and individuals common to a specific domain
